CCDB-39: Object model simple improvements
Task-Url: https://ess-ics.atlassian.net/browse/CCDB-39

Model: removed the constructors and setters for the entity Id. Also
removed the Id as the parameter. This is set by the database
automatically, and can only be read.

Removed all the default contructors (the ones with no parameters) -
actually they are all made private, since they are needed by ther JPA
layer.

In all EJBs and ManagedBeans the new parametrized constructors are used,
mostly with default values, and marked with TODO comments to be
inspected and changed later.